// fct-frontend/src/api/models/add-address-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * REST API for the website of \"Academia Fabiana Perez\"
 * REST API for the website of \"Academia Fabiana Perez\"
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AddressTypeEnum } from './address-type-enum';

/**
 * 
 * @export
 * @interface AddAddressRequest
 */
export interface AddAddressRequest {
    /**
     * 
     * @type {string}
     * @memberof AddAddressRequest
     */
    'street': string;
    /**
     * 
     * @type {string}
     * @memberof AddAddressRequest
     */
    'city': string;
    /**
     * 
     * @type {string}
     * @memberof AddAddressRequest
     */
    'state': string;
    /**
     * 
     * @type {string}
     * @memberof AddAddressRequest
     */
    'zipCode': string;
    /**
     * 
     * @type {string}
     * @memberof AddAddressRequest
     */
    'country': string;
    /**
     * 
     * @type {AddressTypeEnum}
     * @memberof AddAddressRequest
     */
    'addressType': AddressTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof AddAddressRequest
     */
    'userId': number;
}