// fct-frontend/src/api/models/update-address-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * REST API for the website of \"Academia Fabiana Perez\"
 * REST API for the website of \"Academia Fabiana Perez\"
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AddressTypeEnum } from './address-type-enum';

/**
 * 
 * @export
 * @interface UpdateAddressRequest
 */
export interface UpdateAddressRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateAddressRequest
     */
    'street'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateAddressRequest
     */
    'city'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateAddressRequest
     */
    'state'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateAddressRequest
     */
    'zipCode'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateAddressRequest
     */
    'country'?: string;
    /**
     * 
     * @type {AddressTypeEnum}
     * @memberof UpdateAddressRequest
     */
    'addressType'?: AddressTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof UpdateAddressRequest
     */
    'userId'?: number;
}